DCE-92: New file for printing & Add print bitboard to Bitboard Type
