+++ src/engine/evaluation/evaluation.rs
@@ -18,9 +18,7 @@ use crate::engine::misc::bitboard::BitboardTrait;
 use crate::engine::misc::bitboard::Iterator;
 use crate::engine::misc::const_utility::FILE_BITBOARD;
 use crate::engine::misc::const_utility::RANK_BITBOARD;
-use crate::engine::misc::print_utility::print_bitboard;
-use crate::engine::misc::print_utility::print_board;
-use crate::engine::misc::print_utility::print_eval;
+use crate::engine::misc::display::display_board::print_eval;
 use crate::engine::move_generator::bishop::get_bishop_mask;
 use crate::engine::move_generator::bishop::get_bishop_mv;
 use crate::engine::move_generator::bishop::has_bishop_pair;
