+++ src/engine/misc/mod.rs
@@ -2,6 +2,6 @@ pub mod bit_pos_utility;
 pub mod bitboard;
 pub mod const_utility;
 pub mod directions;
+pub mod display;
 pub mod generate_key_utility;
 pub mod play_chess_utility;
-pub mod print_utility;
