+++ src/engine/protocols/uci.rs
@@ -11,7 +11,7 @@ use crate::engine::board::structures::board::Board;
 use crate::engine::board::structures::color::ColorTrait;
 use crate::engine::board::structures::moves::Move;
 use crate::engine::misc::const_utility::FEN_START;
-use crate::engine::misc::print_utility::{from_move_notation, move_notation};
+use crate::engine::misc::display::display_moves::{from_move_notation, move_notation};
 use crate::engine::search::iter_deepening::Search;
 use crate::engine::search::transposition_table::{TTTable, TT};
