+++ src/lib.rs
@@ -41,8 +41,8 @@ pub mod engine {
         pub mod bitboard;
         pub mod const_utility;
         pub mod directions;
+        pub mod display;
         pub mod generate_key_utility;
         pub mod play_chess_utility;
-        pub mod print_utility;
     }
 }
